Move inter-op methods to single class
