--- stdlib/task/src/main/java/org/ballerinalang/stdlib/task/service/Init.java
@@ -1,99 +1,0 @@
-/*
- *  Copyright (c) 2019 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.ballerinalang.stdlib.task.service;
-
-import org.ballerinalang.jvm.util.exceptions.BLangRuntimeException;
-import org.ballerinalang.jvm.values.MapValue;
-import org.ballerinalang.jvm.values.ObjectValue;
-import org.ballerinalang.stdlib.task.exceptions.SchedulingException;
-import org.ballerinalang.stdlib.task.objects.Appointment;
-import org.ballerinalang.stdlib.task.objects.Task;
-import org.ballerinalang.stdlib.task.objects.Timer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Objects;
-
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.FIELD_DELAY;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.FIELD_INTERVAL;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.FIELD_NO_OF_RUNS;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.MEMBER_APPOINTMENT_DETAILS;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.MEMBER_LISTENER_CONFIGURATION;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.NATIVE_DATA_TASK_OBJECT;
-import static org.ballerinalang.stdlib.task.utils.TaskConstants.RECORD_TIMER_CONFIGURATION;
-import static org.ballerinalang.stdlib.task.utils.Utils.getCronExpressionFromAppointmentRecord;
-
-/**
- * Native function to attach a service to the listener.
- *
- * @since 0.995.0
- */
-public class Init {
-
-    private static final Logger LOG = LoggerFactory.getLogger(Init.class);
-
-    @SuppressWarnings("unchecked")
-    public static void init(ObjectValue taskListener) {
-        MapValue<String, Object> configurations = (MapValue<String, Object>) taskListener.get(
-                MEMBER_LISTENER_CONFIGURATION);
-        String configurationTypeName = configurations.getType().getName();
-        Task task;
-        try {
-            if (RECORD_TIMER_CONFIGURATION.equals(configurationTypeName)) {
-                task = processTimer(configurations);
-            } else { // Record type validates at the compile time; Hence we do not need exhaustive validation.
-                task = processAppointment(configurations);
-            }
-            taskListener.addNativeData(NATIVE_DATA_TASK_OBJECT, task);
-        } catch (SchedulingException e) {
-            LOG.error(e.getMessage(), e);
-            //TODO: Ideally this should return an Error using createError() method.
-            // Fix this once we can return errors from interop functions (Check with master)
-            throw new BLangRuntimeException(e.getMessage());
-        }
-    }
-
-    private static Timer processTimer(MapValue<String, Object> configurations) throws SchedulingException {
-        Timer task;
-        long interval = ((Long) configurations.get(FIELD_INTERVAL)).intValue();
-        long delay = ((Long) configurations.get(FIELD_DELAY)).intValue();
-
-        if (Objects.nonNull(configurations.get(FIELD_NO_OF_RUNS))) {
-            long noOfRuns = ((Long) configurations.get(FIELD_NO_OF_RUNS)).intValue();
-            task = new Timer(delay, interval, noOfRuns);
-        } else {
-            task = new Timer(delay, interval);
-        }
-        return task;
-    }
-
-    private static Appointment processAppointment(MapValue<String, Object> configurations) throws SchedulingException {
-        Appointment appointment;
-        Object appointmentDetails = configurations.get(MEMBER_APPOINTMENT_DETAILS);
-        String cronExpression = getCronExpressionFromAppointmentRecord(appointmentDetails);
-
-        if (Objects.nonNull(configurations.get(FIELD_NO_OF_RUNS))) {
-            long noOfRuns = ((Long) configurations.get(FIELD_NO_OF_RUNS)).intValue();
-            appointment = new Appointment(cronExpression, noOfRuns);
-        } else {
-            appointment = new Appointment(cronExpression);
-        }
-        return appointment;
-    }
-}
